requirements.txt
pandas==2.1.4
openai==1.3.7
flask==3.0.0
flask-cors==4.0.0
openpyxl==3.1.2
pyyaml==6.0.1
python-dotenv==1.0.0
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
requests==2.31.0 